Remove any mentions of portability to Hugs.
The `binary` package from which this is derived claims Hugs compatibility, but does not test for this and actual compatibility is not likely except for very ancient historical releases.  As part of cleanup and simplification efforts, the authors of `restore` purposefully make changes that are likely to not be compatible with any compilers that do not target (recent) GHC compatibility.